+++ src/app/modules/appModule.ts
@@ -2,6 +2,7 @@ import { CommonModule, HashLocationStrategy, LocationStrategy } from '@angular/c
 import { HttpClient } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouteReuseStrategy } from '@angular/router';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
@@ -56,6 +57,7 @@ export function HttpLoaderFactory(http: HttpClient) {
 
 @NgModule({
     imports: [
+        BrowserAnimationsModule,
         BrowserModule,
         CommonModule,
         FormsModule,