+++ src/app/widget/openStreetMap/GeoData.ts
@@ -1,0 +1,8 @@
+export interface Point {
+    latitude: number;
+    longitude: number;
+}
+
+export interface Trace {
+    points: Point[];
+}